add new process api to fetch the data from submissions and save to various tables in Civi
